[Python] fix and improve default argument handling
This fixes the incorrect handling of default arguments for Python
for the three new test-suite cases.

Also, this removes the Python 2 vs. Python 3 distinction based on
the `-py3` flag. Now the same python code is produced for default
arguments for Python 2 and Python 3. For this, octal default
arguments, e.g. 0644, are now wrapped as `int('644', 8)`. This is
required, as Python 2 and Python 3 have incompatible syntax for
octal literals.