Use python-config if available in configure
There is no need to check for Python prefix, headers and library ourselves if
we can just ask python-config to do it for us. This won't work for the systems
with very old Python versions, so still keep the existing Python detection
code for them, but use python-config if it's available as it's simpler and
more reliable.

As python-config doesn't return the library directory separately, remove PYLIB
(and PY3LIB, for consistency) and the corresponding make variables and use
just PYLINK which contains both the directory and the library itself instead.