Error thrown from overloaded constructor now throws the specified error.
When an error was thrown from an overloaded constructor, SWIGV8_ESCAPE
would not get called, which is correct. However, the program would then
carry on and get the error message overridden by a general error message
"Illegal arguments for construction of...". Adding an "else goto fail"
skips this line and returns the error that was thrown.