Return enum, not untyped int, when classifying Doxygen commands
This makes the purpose of DoxygenParser::commandBelongs() more clear and also
allows the compiler to check that we handle all of its possible return values.
And it actually turns out that we didn't, so add the missing values now to
avoid changing the behaviour in any way in this commit, but they will be dealt
with later.