Show the original error if calling director method failed
Previously, if calling the director method failed, the precise error
message was lost and just "Error detected" was shown to the user, making
it difficult to understand what the problem actually was.

Change the code to show the contents of the last Python error too to
make finding and fixing errors in Python director methods much simpler.