Terminate options when passed via env var
The C standard requires that argv be terminated with a NULL pointer
(that is, argv[argc] == NULL). There are several places in the swig
codebase that require this to check for missing arguments. However,
SWIG_merge_envopt() was not keeping the NULL terminator, resulting in
argument parsing failures (typically program aborts) when arguments were
passed via the SWIG_FEATURES environment variable.