[Java] Warn when finalizers aren't generated.
A common bug pattern is to %ignoreall and then to forget to %unignore a
destructor. Without a finalizer the delete method isn't called and so
such bugs can be the source of memory leaks without runtime exceptions.
Warn when a finalizer isn't generated so that such memory leaks are easy
to avoid.