Remove long obsolete vms subdirectory
This contains a separate build system for VMS, but it's not been
touched at all since 2002 and clearly hasn't worked for a long time
- e.g. it tries to build php4.cxx, which we removed a decade ago.

It seems openvms is still alive, but keeping these files in the tree
isn't helping anyone. If someone really wanted to build SWIG on
openvms then I suspect starting over would be less work, but they could
pull the files out of the git history if they're really still useful.

Closes #1270.