Resolve types of the classes nested inside template classes
Do the same thing for the template classes as for the normal classes
during the type resolve pass, i.e. call classDeclaration() from
TypePass::templateDeclaration(), in order to fully qualify the names of
nested types used as method parameters inside template classes too.

Also amend classDeclaration() to prevent it from processing base classes
for the templates as this would result in errors in absence of %template
directives for the base templates.

Add a test checking for the situation described above.

Closes #1353.