Use actual parameter names in Python autodoc strings
Since commit fdc6bbeda387fb49ab7dcdbb0f17e4645db3115c Python docstrings
used the auto-generated parameter names of the form "argN" instead of
the real parameter names, making them much less useful.

Fix this by using the actual parameter name (renamed to avoid clashes
with Python keywords if necessary) in the docstring. Unfortunately this
means docstring is now inconsistent with the actual function definition,
which still uses "argN", but this is not as simple to fix and this
change could already be considered an improvement.

See #1272.